It works!!!!!
Removed scanToken() function from the constructor to the main class.
Declaring the Input stream after calling the function means that the scanToken( function will not have that stream. Declaring the input stream before and after means that the input stream is reset after reading the first token.

Added EOF check back into getNextToken(), then when we exercise the
scanner all we have to do is check to see if the next token type is the
EOF token. I have no idea what the lex file is doing to the buffered
reader, but it hates it. I think it reads into its own internal buffer,
thereby exhausting the buffer that is passed into it.